LayoutTests/fast/events/touch/ios/touch-event-listeners-on-subdocuments-expected.txt

Check touch event listeners on sub-documents are correctly cleaned when the document is removed from the document.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



    Tapped (50, 50)
PASS eventList is ['SecondLeveltouchstart']
Tapped (50, 50)
PASS eventList is ['SecondLeveltouchstart', 'FirstLeveltouchstart']
PASS successfullyParsed is true

TEST COMPLETE
